Better keymaps, more general framework.
Keymaps can still not yet be chained, but it is now possible
to use different keymaps.
The keymapped-mixin makes the mixed-in widget listen
to events and call the keymap, and revert to normal behaviour
if no match.

A generic menu allows to customize individual keymaps.
There is one keymap for the canvas, and one for the buttons.
The canvas-keymap has mappings for all buttons, and it is
possible to costumize these mappings.
In particular, one can control-left-click on a button
to edit this mapping.
The control-left-click mapping is itself part of the button-keymap,
and is currently the only mapping in this keymap.
This allows the buttons to have a second callback without changing
the definition of the button.

Some refactoring, code moving is in order.