First command line argument: color-button-size
`racket -l- racket-paint --color-button-size 100`
makes it easier for my 2yo daughter to click on the color buttons.
Needed to create a `create-frame` function to be called after
`command-line`. The function returns a dictionary of the widgets
because I didn't want to `define` all the widgets outside of the
function (because I want them to be available for hacking)
and then set them.